Rename package authentication to security
