+++ eagle-server/src/main/java/org/apache/eagle/server/security/BasicAuthResourceFilterFactory.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.eagle.server.authentication;
+package org.apache.eagle.server.security;
 
 
 import com.sun.jersey.api.model.AbstractMethod;
