+++ eagle-server/src/main/java/org/apache/eagle/server/security/config/AuthenticationConfig.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.eagle.server.authentication.config;
+package org.apache.eagle.server.security.config;
 
 import io.dropwizard.Configuration;
 import com.fasterxml.jackson.annotation.JsonProperty;
