+++ eagle-server/src/test/java/org/apache/eagle/server/security/resource/TestBasicAuthenticationResource.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.eagle.server.authentication.resource;
+package org.apache.eagle.server.security.resource;
 
 import io.dropwizard.auth.Auth;
 import org.apache.eagle.common.security.DenyAll;
